scripts/requirements.txt
pyserial
pyudev

